requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.8
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-registration==2.4.1
django-vote==2.2.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
phonenumbers==8.12.8
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
zipp==3.1.0
